Fixes #24030 - Logic out the SubscriptionReducer
Move the logic out of the SubscriptionReducer.js
into selectors and actions
